pyproject.toml
[tool.poetry]
name = "artsave"
version = "0.1.0"
description = "Telegram bot for exploring Met Museum artworks"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
telethon = "^1.32.1"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 